requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asciitree==0.3.3
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.15.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blosc==1.11.1
cbor2==5.6.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
colorcet==3.1.0
comm==0.2.2
contourpy==1.2.1
cryptography==42.0.7
cuda-python==12.5.0
cycler==0.12.1
Cython==0.29.37
dask==2024.5.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
distinctipy==1.3.4
distributed==2024.5.1
docker==7.1.0
docopt==0.6.2
executing==2.0.1
faiss-cpu==1.8.0
fasteners==0.19
fastjsonschema==2.19.1
fastrlock==0.8.2
figurl==0.2.22
filelock==3.14.0
fonttools==4.51.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.5.0
ghp-import==2.1.0
h11==0.14.0
h5py==3.11.0
hdbscan==0.8.36
hdmf==3.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib_metadata==7.1.0
iniconfig==2.0.0
ipykernel==6.29.4
ipympl==0.9.4
ipython==8.12.3
ipython-genutils==0.2.0
ipywidgets==8.1.2
isoduration==20.11.0
isosplit6==0.1.4
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-highlight-selected-word==0.2.0
jupyter-lsp==2.2.5
jupyter-nbextensions-configurator==0.6.3
jupyter_client==8.6.1
jupyter_contrib_core==0.4.2
jupyter_contrib_nbextensions==0.7.0
jupyter_core==5.7.2
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.2.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.1
jupyterlab_widgets==3.0.10
kachery_cloud==0.4.9
kilosort==4.0.8
kiwisolver==1.4.5
llvmlite==0.42.0
locket==1.0.0
lxml==5.2.2
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mergedeep==1.3.4
mistune==3.0.2
mkdocs==1.6.0
mkdocs-get-deps==0.2.0
mountainsort5==0.5.6
mpi4py==3.1.6
mpmath==1.3.0
msgpack==1.0.8
mtscomp==1.0.2
multidict==6.0.5
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
neo==0.13.1
nest-asyncio==1.6.0
networkx==3.3
notebook==7.2.0
notebook_shim==0.2.4
numba==0.59.1
numcodecs==0.12.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
overrides==7.7.0
packaging==24.0
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
partd==1.4.2
pathspec==0.12.1
patsy==0.5.6
pexpect==4.9.0
phy @ git+https://github.com/cortex-lab/phy.git@8bf8cf751fcbd44210b3c5bbdba8c34a7b2cb1ce
phylib==2.5.0
pickleshare==0.7.5
pillow==10.3.0
platformdirs==4.2.2
pluggy==1.5.0
probeinterface==0.2.21
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pubnub==8.0.0
pure-eval==0.2.2
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.7.2
pydantic_core==2.18.3
Pygments==2.18.0
pykilosort==1.4.6
pynwb==2.7.0
PyOpenGL==3.1.7
pyparsing==3.1.2
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PyQt6==6.7.0
PyQt6-Qt6==6.7.0
PyQt6-sip==13.6.0
pyqtgraph==0.13.7
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.2
pytest==8.2.1
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytz==2024.1
PyYAML==6.0.1
pyyaml_env_tag==0.1
pyzmq==26.0.3
qtconsole==5.5.2
QtPy==2.4.1
quantities==0.15.0
referencing==0.35.1
requests==2.32.2
responses==0.25.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
scikit-learn==1.5.0
scipy==1.13.0
Send2Trash==1.8.3
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sortingview==0.13.3
soupsieve==2.5
spikeinterface==0.100.6
spyking-circus==1.1.0
stack-data==0.6.3
statsmodels==0.14.2
sympy==1.12
tblib==3.0.0
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.3.0
toolz==0.12.1
torch==2.3.0
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
triton==2.3.0
types-python-dateutil==2.9.0.20240316
typing_extensions==4.11.0
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.1
watchdog==4.0.1
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.10
xarray==2024.5.0
yarg==0.1.9
yarl==1.9.4
zarr==2.18.1
zict==3.0.0
zipp==3.18.2
